Create BTCBlock.tcl
This Template overlays Bitcoin blocks as supported by the Bitcoin Core implementation which is the reference implementation. Bitcoin Core is available and currently supported on GitHub. There is no single spec I could find describing the format of the Bitcoin blocks. The format of the block has evolved as Bitcoin Improvement Documents (BIPs) have been adopted. This template supports any implementation that uses the Bitcoin Core scheme for storing the blockchain.