Update BTCBlock-extended.tcl
Minor changes and updated comments. Tested with  10's of thousands of transactions to reveal a few errors. This version is stable but with errors that need to be addressed. Either the parsing interface needs some design rework or  I will backoff the aggressiveness of the parsing.

This is still a work-in-progress and not ready for a Pull request.